requirements.txt
Pillow
requests
tkinter
